+++ test/Information/Externals.jl
@@ -9,6 +9,7 @@ end
 
 if _is_installed("GaussDCA")
 
+    println("Information.gaussdca")
     msa = map(Residue, rand(1:21,100,20))
     dca = gaussdca(msa, min_separation=2)